+++ packages/icons/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 10.5.0 (2024-08-07)
+
 ### New Features
 
 -   Add new `send` icon.